requirements.txt
aiogram~=3.6.0
psycopg2-binary